grouped error messages into modules
